Add pages folder for pages
