+++ src/App.js
@@ -1,18 +1,18 @@
 import "./App.css";
 import { Switch, Route, useLocation } from "react-router-dom";
 import { useEffect, useState } from "react";
-import SignInPage from "./components/SignInPage";
-import BookmarkPage from "./components/BookmarkPage";
-import ResultsPage from "./components/ResultsPage";
-import TakeAwayPage from "./components/TakeAwayPage";
-import OpenPage from "./components/OpenPage";
-import AmbiencePage from "./components/AmbiencePage";
-import CuisinePage from "./components/CuisinePage";
-import ViewPage from "./components/ViewPage";
-import OutdoorPage from "./components/OutdoorPage";
-import PartOfTownPage from "./components/PartOfTownPage";
-import RestaurantDetailedPage from "./components/RestaurantDetailedPage";
-import LoadingPage from "./components/LoadingPage";
+import SignInPage from "./pages/SignInPage";
+import BookmarkPage from "./pages/BookmarkPage";
+import ResultsPage from "./pages/ResultsPage";
+import TakeAwayPage from "./pages/TakeAwayPage";
+import OpenPage from "./pages/OpenPage";
+import AmbiencePage from "./pages/AmbiencePage";
+import CuisinePage from "./pages/CuisinePage";
+import ViewPage from "./pages/ViewPage";
+import OutdoorPage from "./pages/OutdoorPage";
+import PartOfTownPage from "./pages/PartOfTownPage";
+import RestaurantDetailedPage from "./pages/RestaurantDetailedPage";
+import LoadingPage from "./pages/LoadingPage";
 import { parse } from "./utilities/queryString";
 import { takeAwayFilter } from "./utilities/filters/takeAwayFilter";
 import { openingTimesFilter } from "./utilities/filters/openingTimesFilter";
