+++ src/pages/CuisinePage.css
