+++ src/pages/OutdoorPage.css
