+++ src/pages/PartOfTownPage.css
