+++ src/pages/ViewPage.js
@@ -1,32 +1,22 @@
-import "./OpenPage.css";
+import "./ViewPage.css";
 import { ReactComponent as RightArrow } from "../icons/arrow-right-thin.svg";
 import { ReactComponent as LeftArrow } from "../icons/arrow-left-thin.svg";
-import { ReactComponent as Progress } from "../icons/open-progress.svg";
+import { ReactComponent as Progress } from "../icons/view-progress.svg";
 import { Link } from "react-router-dom";
-import FilterTagsGroup from "./FilterTagsGroup";
+import FilterTagsGroup from "../components/FilterTagsGroup";
 
-export default function OpenPage({ restaurantData }) {
-  const tags = [
-    "now",
-    "today",
-    "monday",
-    "tuesday",
-    "wednesday",
-    "thursday",
-    "friday",
-    "saturday",
-    "sunday",
-  ];
+export default function ViewPage({ restaurantData }) {
+  const tags = ["habor view", "park view", "canal view", "alster lake view"];
 
   return (
     <div className="app-grid">
-      <header className="open-header">
-        <h1 className="open-heading">Open</h1>
-        <h2 className="open-subheading">choose one or more</h2>
+      <header className="view-header">
+        <h1 className="view-heading">View</h1>
+        <h2 className="view-subheading">choose one or more</h2>
       </header>
-      <main className="open-main">
-        <div className="open-tag__layout">
-          <FilterTagsGroup filterPage="openingTimes" filterTags={tags} />
+      <main className="view-main">
+        <div className="view-tag__layout">
+          <FilterTagsGroup filterPage="view" filterTags={tags} />
         </div>
         <Link
           to={(location) => {
@@ -36,15 +26,15 @@ export default function OpenPage({ restaurantData }) {
             };
           }}
         >
-          <p className="open-main__amount">{restaurantData.length}</p>
+          <p className="view-main__amount">{restaurantData.length}</p>
         </Link>
       </main>
-      <footer className="open-footer">
+      <footer className="view-footer">
         <Link
           to={(location) => {
             return {
               ...location,
-              pathname: "/take-away-page",
+              pathname: "/cuisine-page",
             };
           }}
         >
@@ -55,7 +45,7 @@ export default function OpenPage({ restaurantData }) {
           to={(location) => {
             return {
               ...location,
-              pathname: "/ambience-page",
+              pathname: "/outdoor-page",
             };
           }}
         >
